requirements.txt
mcstatus
requests
flask
pyyaml